Resolving (most) engine compilation warnings

+++ Common/ac/view.h
@@ -40,7 +40,7 @@ struct ViewFrame {
 
 struct ViewLoopNew
 {
-    short numFrames;
+    int   numFrames;
     int   flags;
     std::vector<ViewFrame> frames;
     // NOTE: we still need numFrames for backward compatibility:
@@ -59,7 +59,7 @@ struct ViewLoopNew
 
 struct ViewStruct
 {
-    short numLoops;
+    int numLoops;
     std::vector<ViewLoopNew> loops;
 
     ViewStruct();
