+++ Engine/ac/room.cpp
@@ -519,7 +519,7 @@ void load_new_room(int newnum, CharacterInfo*forchar) {
     play.bg_anim_delay = play.anim_background_speed;
 
     // Fixup the frame index, in case the new room does not have enough background frames
-    if (play.bg_frame < 0 || play.bg_frame >= thisroom.BgFrameCount)
+    if (play.bg_frame < 0 || static_cast<size_t>(play.bg_frame) >= thisroom.BgFrameCount)
         play.bg_frame = 0;
 
     // do the palette
